+++ packages/lists/src/helpers/index.ts
@@ -1,8 +1,0 @@
-export * from "./general.js";
-export * from "./mute.js";
-export * from "./pin.js";
-export * from "./contacts.js";
-export * from "./bookmark.js";
-export * from "./groups.js";
-
-export * as Operations from "./operations/index.js";