+++ packages/lists/src/queries/index.ts
@@ -1,4 +1,0 @@
-export * from "./mute.js";
-export * from "./pin.js";
-export * from "./bookmark.js";
-export * from "./contacts.js";