+++ packages/user-status/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./helpers/user-status.js";
-export * from "./queries/user-status.js";